Fix the https URL of S3 resources loaded for PFC
In September we changed a bunch of http URLs in our code to https.
Most changes were good as-is, but one related to an S3 resource makes
a Jenkins job fail, because the URL was not designed to be used with https.

In this case, the old URL was:
`http://files.consumerfinance.gov.s3.amazonaws.com/pb/paying_for_college/csv/validated_program_data/[SLUG]`

This was changed to https, but that throws a certificate error when fetching the file from S3.

The https form that we can use (as pointed out by @rosskarchner) is:
`https://s3.amazonaws.com/files.consumerfinance.gov/pb/paying_for_college/csv/validated_program_data/[SLUG]`

Testing
To load an S3 resource locally and confirm that the S3 file is fetched without error, run
```bash
./cfgov/manage.py load_programs --s3 true CFPB_EDMC_20191017.csv
```

You should get an echo saying:
```bash
0 programs created. 2 programs updated.
```