Only deliver video JS to event pages when needed
Event pages are currently loading video-player.js regardless of whether
or not they need it, resulting in a console error if there is not
actually a video module on the page.

This commit makes the loading of that JS conditioned on whether or not
the page is in a state which would have a video module.

Fixes GHE/CFGOV/platform#3569