Add event_state property & context to EventPage
Per comments in #5324, this eliminates the need to import the `when()`
method from `v1.jinja2tags.datetimes` and makes `event_state` available
both as a property of any instance of an `EventPage` and in the context
of any template rendering such.

The `_convert_date` private method from the `v1.jinja2tags.datetimes`
module was moved into a public util module so that it could be accessed
both by the Jinja tag and by the EventPage prop, as well as any future
use.

`start_dt` was made a required field on `EventPage` to simplify the
logic of the `event_state` and `page_js` properties. This necessitated
updates to several unrelated tests that create instances of `EventPage`.

Thanks to @chosak for the guidance!