Add Python test prerequisites to docs
This provides instructions to help people get set up to run the tests. It's fairly straight-forward — either you use the cfgov-refresh standalone setup, or you use the Docker setup, or you install tox (Docker is not currently working because of a pip issue, but I intend to make it work).

It also adds some explanation of our default environments.