Remove PIP_NO_CACHE_DIR
Setting PIP_NO_CACHE_DIR to true breaks our use of venv-update in tox. This means it's not possible to successfully run the Python unit tests in the Docker container.

Our primary use of Docker at the moment is for local development, so it seems prudent to preserve the ability to do all aspects of development, including run the test. The cost of longer build time and some warnings seems worth that to me.

We should revisit this when/if we deploy Docker containers for deployment.