Bypass CloudFront when accessing HUD search results in S3
When there was an error fetching a file from the bucket (a 500 error
in S3), CloudFront would cache the error for five minutes, meaning
that even if the underlying error was resolved, end users would still
receive an error when trying to visit that page in the FAHC tool.

We have code to retry a request to S3 if it receives a 5XX error, which
was not useful when CloudFront was caching responses. By bypassing CF,
requests that 500 will be retried immediately, which may reduce the
number of user-facing 500 errors.