Add unique slugs to expandable group headings
This change generalizes some behavior that was introduced in PR 5334,
adding an 'id' attribute to the div wrapping ExpandableGroups, based on
a slugified version of the EG's heading (if specified).

This makes it easier to generate in-page links to ExpandableGroups, and
is now turned on by default for all existing EGs.

A new slugify_unique Jinja filter has been created to support this
behavior.

To test, run a local server, create a page that supports EGs (like a
LearnPage) and add one or more. Inspect page source to see the generated
'id' attribute. Note that if you generate multiple EGs with the same
heading, their 'id' attributes will be different.