Restore content on print (#5363)
The stylesheet for the forthcoming YES car buying guide was a little too 
aggressive in hiding things on print, and it was being applied globally. 
Since that page hasn't launched, I have commented it out of the global 
`main.less` file to fix the immediate problem. Further consideration 
will need to be given to how to scope the styles that the car buying 
guide will need to only that page.