Add helper script to sync local image storage
When developing locally using a production database dump, it can be
distracting and difficult to work with Wagtail images. Images (and their
renditions) are stored in the database, but in production the image
files themselves are stored in S3. When developing locally with a dump,
you have the references to those files, but not the files themselves.
This means missing images when browsing pages, which can break layout
and make it hard to debug and develop visual styles.

This commit adds a new Django management command, sync_image_storage,
that can be used to download in bulk all images and renditions
referenced in the database.

To run:

cfgov/manage.py sync_image_storage https://files.consumerfinance.gov/f/ ./cfgov/f/

See documentation change for a similar explanation.

Note that this can take a little while as it downloads quite a few files
(currently 4000+ files, ~400MB).