Add hardcoded InfoUnits to new homepage design
This commit adds some hardcoded InfoUnits to the new homepage design,
using the 33/33/33 InfoUnitGroup template. This content is hardcoded
pending implementation as editable Wagtail blocks. Content is
placeholder and is the same in both English and Spanish right now.

To test, visit the English and Spanish versions of the new homepage:

- http://localhost:8000/?nhp=True
- http://localhost:8000/es/?nhp=True