Update carousel image placeholder and logic
This commit updates how the new home page carousel images get added to
the page. Instead of hardcoding a URL, we want to use a Wagtail image.

This change makes incremental progress towards this goal, updating a
hardcoded URL to be a hardcoded Wagtail image ID. It also makes use of
the built-in ability to generate renditions and uses the 'fill-50x50'
filter to create the thumbnail images.