Add Spanish translation for "Featured" on homepage
The new homepage design includes the word "Featured" above the carousel
when viewed at narrow screen widths (mobile). We want this word to
translate properly for Spanish versions of the page.

To test, run a local server and visit http://localhost:8000/es/?nhp=True
at a mobile width. You should see the translation "Destacado". This
comes from the existing translations for this project:

https://github.com/cfpb/cfgov-refresh/blob/aaf5405c55aef08f424f86d4a63b9736f26ccaf7/cfgov/locale/es/LC_MESSAGES/django.po#L78