Migrate new home page content into Wagtail
This set of changes moves the content for the new home page design
into configuration for the Wagtail HomePage page type.

Serving the new content from hardcoded values is still supported via the
existing NEW_HOME_PAGE feature flag. A new NEW_HOME_PAGE_IN_WAGTAIL
feature flag can be used to instead serve the content from Wagtail. If
both flags are enabled, the Wagtail one takes precedence.

There are several changes here related to Wagtail InlinePanels and the
django-modelcluster package that powers them. In the version of Wagtail
that we are using (1.13), nested InlinePanels are not supported at all.
In versions >2.0, nested InlinePanels are supported via a workaround
detailed on wagtail#5511. Full support for nested InlinePanels will be
available in Wagtail starting in version 2.8, see wagtail#5566.

It turns out that we can use the nested InlinePanel workaround on
Wagtail 1.13, but to do so we need to increment the version of
django-modelcluster from 3.1 to 4.4. Unfortunately, Wagtail 1.13
requires django-modelcluster<4.0 even though 4.4 still supports it,
along with Python 2 (support for both of these is not removed until
django-modelcluster 5.0). So we want to force use of
django-modelcluster==4.4, even though it technically conflicts.
Unfortunately, our use of tox_pip_extensions_ext_venv_update prevents
this (it doesn't allow you to specify any conflicting versions, even
though pip does out of the box; see venv-update#105 on this including a
comment from an old colleague). So, I've (hopefully temporarily)
disabled that until we are on Wagtail 2.

After all that you may be wondering why use nested InlinePanels at all.
The alternative to this would be to use StreamFields, but this felt like
a worse implementation to me: harder migrations, worse user experience
in the Wagtail editor, etc. I welcome any feedback on this choice.

These changes include a new Django management command to automatically
populate the Wagtail fields from the hardcoded values. After running
the new v1 migration on your database, run

cfgov/manage.py migrate_home_page_content

to update both the English and Spanish pages.