Use custom MEDIA_ROOT for Python tests
Some of our Python tests involve creating Django media files, for
example tests that exercise Wagtail image upload. These files get stored
in settings.MEDIA_ROOT which, for local development, is the cfgov/f/
directory. This directory thus gets filled over time with numerous image
files named test_something.png, with new ones being created every time
the tests are run.

It would be better if the tests instead used some other MEDIA_ROOT, so
that the directory used for local development isn't altered by the
tests, and so that the tests clean up after themselves. This commit
makes such a change.

With this change, the test MEDIA_ROOT is set to cfgov/tests/test-media.
This directory is removed after the tests are run. This pattern is
inspired by how Wagtail does this, see

https://github.com/wagtail/wagtail/blob/96c1c41daa391ebe3cc3daa102f354ca363c34be/wagtail/tests/settings.py#L6
https://github.com/wagtail/wagtail/blob/b3f1f50fba8576fc80368a6a402b49f16771a4fe/conftest.py#L48

A good way to test this is to first look at your cfgov/f/ directory, and
notice how, surprisingly, it is full of files that you don't remember
uploading. Then, remove it, and run tests on master. Notice that files
get written there. Then repeat the process on this branch, and observe
nothing gets left behind.