Move the GlossaryTerm model to ask_cfpb
This is an attempt to resolve a circular dependency we have in our migrations for ask_cfpb and v1. The GlossaryTerm model is closely tied to Ask CFPB per @higs4281, so I've moved it to the ask_cfpb app. This is accomplished with a migration that alters the table name in the database and then "deletes" the v1.GlossaryTerm model for the purposes of the Django migration state. Another migration then adds the ask_cfpb.GlossaryTerm model to the Django migration state without performing any database actions.

I've left the v1.wagtail_hooks alone, which adds GlossaryTerm to the snippets menu. The GlossaryTermAdmin is still defined there.