Fix the delivery of answer summaries for search-by-tag pages. (#5445)
The search-results template for Ask CFPB was tasked with handling both
Elasticsearch results and results from a dark corner of the Ask universe,
search-by-tags. The template wasn't handling answer streamfield content
correctly, and this surfaced when an answer showed up that had non-text content,
specifically a video player.

This change simplifies the template logic and moves the summary extraction
into the page logic, re-using the methods from our search indexing.