Treat `filter()` object as iterable
This change fixes the Ask CFPB data export that was broken in Python 3 because of a difference in the way the `filter()` function works.

`filter()` in Python 2 returns a list, and so could be called with list indices. `filter()` in Python 3 returns an iterable `filter` object that cannot be treated like a list.

Since we want to first object resulting from the filter, the fix is a single call to `next()` in place of `[0]`.