Fix job listings with multiple USAJobs links
This is a fix for job listings that have multiple USAJobs links in them. With Python 3's stricter comparisons, attempting to order the `USAJobsApplicationLink` objects using `applicant_type` was failing with `ApplicantType` not supporting `<` operations.

This fix adds `__gt__` and `__lt__` to `ApplicantType` to support `>` and `<` operations.

Generally this appears to come from django-modelcluster and behavior that differs from the Django expectation, possibly related to https://github.com/wagtail/django-modelcluster/issues/45.