Prep imports and code for Wagtail 2.3
Wagtail 2 introduced breaking changes because the core Wagtail packages
were renamed. Upgrading will require two steps for us:
1. Add conditional imports that handle either Wagtail 1.13 or 2.3, for testing.
2. Jump to Wagtail 2.3 and remove the conditional code.

This is step 1.

We used two patterns for conditional imports: try/except blocks, and, in
migrations, if/else blocks.

In addition to imports, this PR also will make these changes:
- Preserve a Wagtail function (wrap_password_reset.py) that was removed
in Wagtail 2, but is needed by our custom handling of password resets.
- Rework /v1/management/commands/update_data_snapshot_values.py, which
wouldn't run in Wagtail 2 because of the way it stored page objects.
- Adjust cfgov/cfgov/test.py to avoid duplicating initial data when
using -fast and -slow tox tests. We may want to revisit the value of having
the fast and slow configuration, since we squashed migrations recently.
- Fix flake8 and isort errors.