Better handle header layout with no mega menu
If no mega menus have been defined in Wagtail, the new mega menu
implementation (introduced in PR 5471 via the MEGA_MENU_BACKEND_V2
feature flag, and now turned on in production) will raise an exception.
This is problematic because it prevents us from running Python tests
with defining some placeholder menu, and prevents use of cfgov-refresh
on an empty database without defining a menu.

To better handle these cases, it would be better if we could instead
just fallback to showing no menu if none are defined. Additionally, this
opens up the possibility of deliberately hiding the menu on certain
pages in an organized way, as opposed to via custom base templates.

This change makes that possible, and makes the necessary frontend code
changes to handle the case where the menu is not displayed.