Fix import-sorting in paying_for_college
When we incorporated the paying_for_college satellite app into cfgov-refresh,
we neglected to add a reference in our .isort.cfg file. The result is that
isort doesn't group paying_for_college with other local cfgov-refresh apps.

This PR brings the app into the fold and resorts imports.
No changes in functionality are expected.