Remove extra borders from FCMs on some page types (#5528)
* Remove extra borders from FCMs

The Featured Content Module component has provided its own border for a 
long time. We apparently haven't noticed that it gets doubled up on 
Browse, Browse Filterable, and Newsroom Pages, probably because FCMs are 
rarely used on those pages types. This commit removes the class in the 
markup for those page type templates that adds the extra border.

* Also remove extra border on future Choose a Student Loan template