Add basic `watch` gulp task
This is a proof of concept that adds a basic watch task for uncompiled Less and JS during local development. It watches files in the src directories for changes and runs the 'modern' gulp tasks when they change. This will not run commands to update the application specific styles/scripts, the on demenand atomic components, or legacy browser support.