Support editing of raw stream data
The get_stream_data function attempts to return the raw JSONish
StreamField data, but does so only after the block value is first
converted to Python. This means that if the block's schema has changed
(for example, a field has been removed in a previous AlterField step),
those fields will be removed from the raw data.

Instead of going through the Python conversion and letting Wagtail strip
out any blocks that it doesn't think should be there, we want to access
the raw DB content exactly as it is stored.