Remove mega menu side borders on mobile
Setting left and right borders on a nested transform3d() transitioned
element results in the element's width being incorrectly calculated. I'm
not sure why this is but for us it results in a gap on the right of the
mega menu's second level for all mobile devices.

GHE/CFPB/el-camino/issues/142