Add new EnforcementActionPage with metadata (#5523)
Currently, we use the Document Detail page type to create pages about
Enforcement Actions in Wagtail, and use the Related Metadata streamfield
block to store important information about the action. This means that
the metadata (which should be present on all enforcement actions) might
be missing or malformed, and is not easily searchable or filterable.

This PR adds an Enforcement Action page type, which includes a new tab
in the wagtail interface called "Metadata". It contains database fields
for the required metadata. This will allow us to filter on these values
in the Enforcement Actions filtered list page in a future update.

This includes a new model, EnforcementActionStatus, since a single
enforcement action can be taken against multiple institutions and might
have a different status with each.

The migration to add the new models is written to be compatible with
both Wagtail 1.13 and Wagtail 2.3, since we are in the midst of that
upgrade.