Mega menu: cleanup of obsolete code
With the move to the new mega menu, we have some code that can be
cleaned up. Specifically, the new menu doesn't support featured content
with images or bodies. Additionally, some of the data fields were named
a certain way to support compatibility with the old menu; here I've
renamed them to be more consistent ("url" vs. "external_link", for
example).