Remove pinned django-modelcluster requirement
Now that we're on Wagtail 2, we don't need to pin django-modelcluster to
version 4+. This pin was only needed when we were still on Wagtail 1.13.

This removes one of the inconsistent packages that was preventing the
restoration of tox extensions; however there are now two other ones that
need to be resolved.