Simplify job posting StreamField blocks
This PR refactors and simplifies the StreamField blocks used to show a
list of job postings on the site. Currently there are two separate
blocks for this, a JobListingList (displayed on the careers overview
page [0]) and the JobListingTable (displayed on the current openings
page [1]).

I implemented these blocks way back in 2016 (PR 2314) as part of my
first development with Wagtail and in retrospect think I overengineered
these. Instead of making specific block types as needed for the job
listings, I created some generic block types (ModelList and ModelTable)
and also exposed quite a few configuration options for job listings
which historically have never been used once.

Since their creation, these block types have only ever existed in a
single place and in a single configuration. In the interest of
simplification, then, this PR removes a whole lot of complexity and
tries to simplify to only the use case we need.

I believe that ModelList was temporarily used at one point for some Ask
CFPB functionality, but I believe that never made it to production.

This PR does leave behind the generic ModelBlock block, also created
around the same time. It'd be nice to clean this up as well, but it's
currently being used by RegulationsList [2]. I note that this is
replaced by a more specific BaseRegulationsList block type [3] in the
hopefully-soon-to-be-used replacement wagtail-regulations package [4],
so hopefully we can do that in future.

The best way to test these changes is to visit the local versions of
these pages, and compare them against the production versions:

http://localhost:8000/about-us/careers/
http://localhost:8000/about-us/careers/current-openings/

Besides that, the unit tests have been refactored and still provide 100%
coverage of these block types.

For additional context, this work is being done to smooth some upcoming
changes to add multiple locations to job postings.

[0] https://www.consumerfinance.gov/about-us/careers/
[1] https://www.consumerfinance.gov/about-us/careers/current-openings/
[2] https://github.com/cfpb/cfgov-refresh/blob/1b478eb9fe2b61ea29a344e898028da4f92130eb/cfgov/regulations3k/blocks.py#L11
[3] https://github.com/cfpb/wagtail-regulations/blob/cb515e8b9c0e6a16a667cdf81e544d4973f5373d/wagtailregulations/blocks.py#L16
[4] https://github.com/cfpb/wagtail-regulations