Update the ping_edmc script to help test outgoing IPs
When we set up notifications for EDMC, the first settlement school
to use our 'EFIP' paying-for-college disclosure tool, we used the ping_edmc.py
script to test the notification routine end-to-end.

One crucial bit that we needed to test was a list of IPs that our web
configuration might use to send notifications to EDMC's endpoint.
The school planned to whitelist our IPs so that we were the only
entity that could post to their endpoints.

Our outbound IPs will be changing soon, and we'll need a way to test
the new IPs. This PR adjusts the script to make that easier and renames it
to make clear that it is no longer EDMC-specific.

I also took this chance to get the latest school endpoints from our
database rather than from a hard-coded variable.

For context, see documentation at
[GHE]/paying-for-college/dynamic-disclosures-documentation/wiki/2020-Notification-changes