Add global banner manager to Wagtail
We have created a lot of banners that live above everything else, at the 
very top of the page, with a 
[complex](https://github.com/cfpb/cfgov-refresh/blob/master/cfgov/jinja2/v1/_includes/organisms/header.html#L28-L61) 
[flag-based](https://github.com/cfpb/cfgov-refresh/blob/6de3110eac363de2969043fbe7c8ccb51f50632b/cfgov/cfgov/settings/base.py#L631-L656) 
[apparatus](https://github.com/search?q=repo%3Acfpb%2Fcfgov-refresh+extension%3Ahtml+filename%3Abanner+language%3AHTML+language%3AHTML&type=Code&ref=advsearch&l=HTML&l=HTML). 
This requires developer intervention for every new banner that gets 
created, and it is piling on a lot of conditional code and tech debt 
that could better be handled by storing some of this configuration in 
the database.

This commit creates a new Wagtail-based banner manager to do just that.
A new Banner Django model is added, along with a ModelAdmin interface 
for creating and editing Banner instances.

Banners have a title used for internal reference, a URL pattern field 
for determining what paths it appears on, and instance of the 
Notification molecule (providing the visible content of the 
notification), and a checkbox to manually enable/disable the banner.

See: https://GHE/CFGOV/platform/issues/3677