Optionally display Contact body in expandables (#5610)
Currently when Contacts are show in expandables (as in the Contact Us
page), we never show the Contact snippet body. We want to add the
ability to do this on a snippet-by-snippet basis.

This change adds a new body_shown_in_expandables boolean to Contact
snippets, that is used to determine whether the body is shown.