Add CampaignPage
There is a desire for a page type that has a sidebar-less template with 
a larger hero and an alternate features box, similar to our homepage 
design from Spring 2016 – Fall 2019. This commit adds a new 
`CampaignPage` page type to do that, supported by a new option on the 
`Hero` molecule to render a "jumbo" hero.

See:
- https://GHE/CFGOV/platform/issues/3573#issuecomment-231301
- https://GHE/CFPB/Consumer-Facing-Iteration/issues/57