Ensure we have a copy of gh-pages locally
In our docs action we did not have a copy of the gh-pages branch, so one was being created. When mkdocs tried to push it, it was rejected because there were changes on gh-pages upstream that we didn't have. This change will fetch the last commit on gh-pages to prevent that from happening.