Specify Postgres as the default test database
We use a postgres-specific Django field, JSONField, in several tables,
and to test this feature we can't use sqlite. So we need to specify
our test database as Postgres, rather than leaving that optional.

This adds an uncommented line to our `.env_SAMPLE` file to specify Postgres
for local development and testing.

Documentation is updated to note that we're dropping support for sqlite and
to change setups so that Django can create test Postgres databases.