Default to using PostgreSQL for tox
This change will make tox default to using a local Postgres if `TEST_DATABASE_URL` is not otherwise defined.

This will have the result of defaulting to the Postgres Docker container with the previous commit. This change also brings the GitHub Actions Postgres container in line with our local Docker container, so it can use the same (default) value of TEST_DATABASE_URL.

The end result is that just running `tox` while the Docker containers are running will use Postgres.