Give iRegs pages better titles and descriptions (#5625)
Improves regulation page titles and descriptions

Interactive Bureau Regulation pages all had the default cf.gov meta description
and a title that was just the name of the regulation. This changes the titles
of all the pages to match whatever the page heading is and the meta and
Open Graph descriptions to a summary of the page that's unique across all
the pages (better for SEO!).