Add CampaignPage page type with new JumboHero block (#5614)
* Add CampaignPage

There is a desire for a page type that has a sidebar-less template with 
a larger hero and an alternate features box, similar to our homepage 
design from Spring 2016 – Fall 2019. This commit adds a new 
`CampaignPage` page type to do that, supported by a new option on the 
`Hero` molecule to render a "jumbo" hero.

See:
- https://GHE/CFGOV/platform/issues/3573#issuecomment-231301
- https://GHE/CFPB/Consumer-Facing-Iteration/issues/57

* Fix IUG template when value.intro is None

* Remove unnecessary imports copied in

* Add tests for richtext_isempty filter

* Fix isort issues

* Update richtext_isempty filter

* Refactor jumbo hero into its own block, rebuild migrations

Thanks to the suggestion of the illustrious @chosak (and finding 
ourselves with a bit of deadline breathing room), I undertook the 
separating out of the Jumbo Hero into its own block, to avoid needing to 
add the extra is_jumbo option to the existing Hero block. Both regular 
and jumbo hero blocks now inherit their shared fields from a new 
AbstractHero class.

* Alternate approach to JumboHero logic

* Also look for rich text values with nothing but whitespace

Co-authored-by: william higgins <higs4281@users.noreply.github.com>
Co-authored-by: Andy Chosak <andy.chosak@cfpb.gov>