Handle questions with no answers in Ask CFPB export
Our Ask CFPB CSV export expected all questions to have an answer of some kind, be it a text block, or FAQ, or How To. If a question is created as a draft, but does not have any answer text yet, the export will fail.

This change fixes that issue.