Strip current path from in-page anchor links
This is a temporary way to work around the fact that Wagtail doesn't
include the ability to insert an anchor link with an `href` of just
`#anchor` until Wagtail 2.7. Currently, one must input the root-relative
path to the current page before the fragment identifier.

We'll be on 2.8 soon enough and have access to the new anchor link
feature, but the need for this is pressing, to allow print CSS to target
links beginning with `#` in order to unstyle them on the print version
of a page.