Update old-style middleware
This change updates our middleware from the old pre-Django 1.10 style to the new post-Django 1.10 style. As part of this we'll be using `MIDDLEWARE` instead of `MIDDLEWARE_CLASSES` on Django 1.11 as well as 2.0.

This fixes middleware errors when running the tests with Django 2.0.

It also adds tests for `DownstreamCacheControlMiddleware`.

This is the result of pairing that @cwdavies and I did.