Use Postgres as the default test database (#5627)
* Specify Postgres as the default test database

We use a postgres-specific Django field, JSONField, in several tables,
and to test this feature we can't use sqlite. So we need to specify
our test database as Postgres, rather than leaving that optional.

This adds an uncommented line to our `.env_SAMPLE` file to specify Postgres
for local development and testing.

Documentation is updated to note that we're dropping support for sqlite and
to change setups so that Django can create test Postgres databases.

* Open Postgres container port for use locally

This change opens up the Postgres container's PostgresSQL port to use from the host.

* Default to using PostgreSQL for tox

This change will make tox default to using a local Postgres if `TEST_DATABASE_URL` is not otherwise defined.

This will have the result of defaulting to the Postgres Docker container with the previous commit. This change also brings the GitHub Actions Postgres container in line with our local Docker container, so it can use the same (default) value of TEST_DATABASE_URL.

The end result is that just running `tox` while the Docker containers are running will use Postgres.

* remove settings config for sqlite test databases

* comment out the TEST_DATABASE_URL var in .env

* Fix import sorting in settings/test.py

* Update .github/workflows/backend.yml

Co-Authored-By: Will Barton <william.barton@cfpb.gov>

* Fix an outdated reference to the 'python3' container, and some rewording

* remove references to TEST_DATABASE_URL in tox.ini

* Remove configs for TEST_DATABASE_URL from settings/test.py

* Remove TEST_DATABASE_URL references in .env_SAMPLE

* Remove reference to TEST_DATABASE_URL from .travis.yml

Co-authored-by: Will Barton <william.barton@cfpb.gov>