Simplify anchor link handling in parse links middleware
Prior to this change we were using `request.path` as a part of a regular expression compilation. This depends on the request path being a valid regular expression. We encountered an issue where a request path in production was not.

This change, suggested by @wpears, does a straight string comparison between an link href and the request path looking for an anchor's `#`.

This causes the bad request path we saw in production (`/about-us/blog/worried-about-making-your-auto-loan-payments-your-lender-may-have-options-to-help/%20)`) to 404 instead of 500.