Ask CFPB: Remove extra newlines from search index
Currently when we index Ask CFPB AnswerPages into Elasticsearch, both
the URL and autocomplete text have extra newlines at the end due to the
use of templates. This doesn't affect the functionality of the site in
practice, but can be observed by inspecting the HTML on certain Ask
pages, and would be good to fix for purposes of index cleanliness.

To test, if you query an example document using the master branch,
you'll see something like this:

$ curl "http://localhost:9200/content_haystack/_search?size=1&fields=autocomplete,url&pretty=true"
{
  ...
  "hits" : {
    ...
    "hits" : [ {
      "_index" : "content_haystack",
      "_type" : "modelresult",
      "_id" : "ask_cfpb.answerpage.5782",
      "_score" : 1.0,
      "fields" : {
        "autocomplete" : [ "What should I do if my EBT card or PIN is lost or stolen or I see unauthorized charges?\n" ],
        "url" : [ "/ask-cfpb/what-should-i-do-if-my-ebt-card-or-pin-is-lost-or-stolen-or-i-see-unauthorized-charges-en-2055/\n" ]
      }
    } ]
  }
}

Note the extra newlines at the end of the autocomplete and URL fields.

After this change and a reindex, those are correctly removed:

$ cfgov/manage.py update_index ask_cfpb
$ curl "http://localhost:9200/content_haystack/_search?size=1&fields=autocomplete,url&pretty=true"
{
  ...
  "hits" : {
    ...
    "hits" : [ {
      "_index" : "content_haystack",
      "_type" : "modelresult",
      "_id" : "ask_cfpb.answerpage.5782",
      "_score" : 1.0,
      "fields" : {
        "autocomplete" : [ "What should I do if my EBT card or PIN is lost or stolen or I see unauthorized charges?" ],
        "url" : [ "/ask-cfpb/what-should-i-do-if-my-ebt-card-or-pin-is-lost-or-stolen-or-i-see-unauthorized-charges-en-2055/" ]
      }
    } ]
  }
}