Change the final estimation paragraph to avoid use of "certificate"
The logic in the last section of the EFIP disclosure tool is failing and always
calls the program's document a certificate, regardless of the degree earned.

This can be misleading, and we need a hot fix to remove the reference to certificate.

A longer-term solution will be to rewrite js code to pick up program level information.