Fix: Ignore request headers for mega menu language
Currently it's possible for request headers to impact the language that
the mega menu displays in. Specifically, setting the
HTTP_ACCEPT_LANGUAGE header to Spanish could cause the menu to appear in
Spanish on certain English pages using a legacy base template.

To test, update your browser settings to prefer Spanish (e.g. in Chrome,
Settings, Advanced, Languages, Language) and visit

http://localhost:8000/data-research/consumer-complaints/

On master, the mega menu will be in Spanish. On this branch, it'll be in
English.

! Note that other parts of the page may still incorrectly be in Spanish
-- this is a separate issue unrelated to the mega menu code.