Fix the iRegs wayfinder
The iRegs wayfinder broke when we started adding u-hidden classes to collapsed
expandable content. The u-hidden class gives those elements a display: none,
which means their positions can't be correctly calculated for the map of all
paragraphs on the page that powers the wayfinder. This change fixes that by
toggling the u-hidden class off while the map of paragraph positions is being
constructed, then toggling it back on again after the map is done.