Fix our HowTo schema markup support per newest guidance
At some point since we first implemented it, Google made the `name` 
field required for the `HowTo` itself (as opposed to just the individual 
`HowToStep`s). This commit adds a new corresponding `title` field to the 
`HowTo` block in Wagtail, and outputs that in the template.

Also includes some other minor cleanup of that template, including 
putting the steps inside an ordered list.

See Google's latest HowTo guidance at: 
https://developers.google.com/search/docs/data-types/how-to