Remove PyYAML from libraries and update it in docs
This upgrade is to address CVE-2017-18342. As of #5714 it's no longer required in libraries.txt.

There was a vulnerability in PyYAML 3.1.13 that we determined didn't effect us in the way we use the library, but the vulnerability still gets flagged. Previously we couldn't upgrade PyYAML because non-vulnerable versions did not support Python 2. Now we can!