Remove attempts to resolve modified path
Discovered that `resolve` will always return a successful result because 
any URLs that don't match a standard pattern match the Wagtail fallback 
pattern. Didn't catch this before because a bug in the slash-appending 
logic was causing a slash to always be appended, which was "correctly" 
failing to resolve some URLs.

@chosak and I agreed that it wasn't worth the complexity to test for 
both Django and Wagtail URLs (which would involve getting the current 
site and testing with one of its class methods), so falling back to just 
doing a redirect if any transformation of the URL occurred.