Downgrade wagtail-autocomplete
wagtail-autocomplete 0.5's README.st is UTF-8 encoded, and somewhere along our build pipeline the README.st is not being opened UTF-8 encoded (I believe because the locale is not set to UTF-8 in a way that it needs to be). This downgrades to unbreak the build while we figure that out.