Add ability to force breadcrumbs for homepage children
This change adds a checkbox to all CFGOVPage pages that, if checked when they are direct children of the home page, will cause their children to include that page in its breadcrumbs.

This is necessary for some top-level pages that are not available through the mega menu.

This change also includes some optimization as well, thanks to @chosak!